moved i18n and validate setting into seperate files
