+++ frontend/src/locales/da.js
