+++ frontend/src/locales/da/alerts.json
