+++ frontend/src/locales/da/buttons.json
