+++ frontend/src/locales/da/common.json
