+++ frontend/src/locales/da/input_fields.json
