+++ frontend/src/locales/da/navbar.json
