+++ frontend/src/locales/da/shared.json
