+++ frontend/src/locales/da/table_headers.json
