+++ frontend/src/locales/da/tabs.json
