+++ frontend/src/locales/da/workflows.json
